requirements.txt
# Core ML Libraries
ludwig==0.10.4
mlflow==2.8.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Data Processing
pyarrow==14.0.1
dask==2023.11.0

# Model Serving (compatible with Ludwig's pydantic<2.0 requirement)
fastapi==0.95.2
uvicorn==0.24.0
pydantic>=1.10.0,<2.0.0

# Containerization & Deployment
docker==6.1.3
kubernetes==28.1.0

# Monitoring & Logging
prometheus-client==0.19.0
grafana-api==1.0.3
loguru==0.7.2

# Configuration Management
pyyaml>=5.0.0,<6.0.1,!=5.4.*
python-dotenv==1.0.0
hydra-core==1.3.2

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Visualization (for model registry UI)
plotly>=5.0.0,<6.0.0
streamlit>=1.28.0,<2.0.0
